pyproject.toml
[project]
name = "treescan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.0.0",
    "pyyaml>=6.0.3",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
